+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add `@ApiProperty`.type for `@DtoCastType'
 - support nested parentheses for validators, e.g.: @Matches(/^[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}$/i)
 - add annotation `@CustomValidator`
+- always set `@ApiProperty`.required
 
 ## [1.17.4] - 2023-03-04